+++ src/background_scripts/extractors/kissmanga.js
@@ -24,15 +24,26 @@ export function getChapterReference(url, defaultValue) {
  * @returns String representing the manga cover URL.
  */
 export function getMangaCover(response, url) {
-  if (!response || typeof response !== 'object') {
-    throw FoxyError(ErrorCode.RESPONSE_NOT_HTML, url);
-  }
-
-  // Get manga image URL from response
   const imageDom = response.evaluate('//link[@rel="image_src"]', response, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null);
   if (!imageDom.singleNodeValue) throw FoxyError(ErrorCode.NO_MANGA_COVER, url);
 
   return imageDom.singleNodeValue.getAttribute('href');
+}
+
+/**
+ * Returns the Manga status (Ongoing|Completed) from the manga main page.
+ * @param {object} response The Fetch API response object.
+ * @param {string} url Optional. The URL to print with errors.
+ * @returns {boolean} True if the series is stated as 'Completed' and False otherwise.
+ */
+export function getMangaStatus(response, url) {
+  const imageDom = response.evaluate('//div[@id="leftside"]//div[@class="barContent"]/div[2]', response, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null);
+  if (!imageDom.singleNodeValue) throw FoxyError(ErrorCode.NO_MANGA_STATUS, url);
+
+  const statusStr = /Status:\s(\w+)/.exec(imageDom.singleNodeValue.textContent);
+  if (!statusStr) throw FoxyError(ErrorCode.NO_MANGA_STATUS, url);
+
+  return (statusStr[1] === 'Completed');
 }
 
 /**
@@ -113,8 +124,9 @@ export function getMangaInfo(url) {
 
     const name = titleDom.singleNodeValue.getAttribute('content').replace(/Read (.+) manga .+/, '$1');
 
-    // Get manga SID and image URL from response
+    // Get image URL and status from response
     const imageUrl = getMangaCover(response, mangaUrl);
+    const status = getMangaStatus(response, mangaUrl);
 
     // Get chapter list and return
     try {
@@ -128,6 +140,7 @@ export function getMangaInfo(url) {
         reference: mangaReference,
         url: mangaUrl,
         cover: imageUrl,
+        status,
         last_update: new Date(),
         chapter_list: chapterList,
       };
@@ -161,4 +174,28 @@ export async function updateChapters(manga) {
   } catch (err) {
     throw err;
   }
+}
+
+/**
+ * Returns the manga information
+ * @param {object} manga The manga to update.
+ * @returns {Promise} A promise which resolves to the manga object retrieved.
+ */
+export async function updateMetadata(manga) {
+  if (!manga || !manga.url) {
+    throw new TypeError(`updateMetadata() argument is invalid: ${JSON.stringify(manga)}`);
+  }
+
+  return HttpFetch(manga.url, async (response) => {
+    if (!response || typeof response !== 'object') {
+      throw FoxyError(ErrorCode.RESPONSE_NOT_HTML, manga.url);
+    }
+
+    const update = {
+      cover: getMangaCover(response, manga.url),
+      status: getMangaStatus(response, manga.url),
+    };
+
+    return update;
+  });
 }