+++ src/background_scripts/extractors/mangaeden.js
@@ -24,10 +24,6 @@ export function getChapterReference(url, defaultValue) {
  * @returns String representing the manga cover URL.
  */
 export function getMangaCover(response, url) {
-  if (!response || typeof response !== 'object') {
-    throw FoxyError(ErrorCode.RESPONSE_NOT_HTML, url);
-  }
-
   // Get manga image URL from response
   const imageDomList = response.evaluate('//meta[@property="og:image"]', response, null, XPathResult.ORDERED_NODE_SNAPSHOT_TYPE, null);
   if (imageDomList.snapshotLength <= 0) throw FoxyError(ErrorCode.NO_MANGA_COVER, url);
@@ -36,6 +32,22 @@ export function getMangaCover(response, url) {
     ? imageDomList.snapshotItem(1).getAttribute('content') : imageDomList.snapshotItem(0).getAttribute('content');
 
   return imageUrl;
+}
+
+/**
+ * Returns the Manga status (Ongoing|Completed) from the manga main page.
+ * @param {object} response The Fetch API response object.
+ * @param {string} url Optional. The URL to print with errors.
+ * @returns {boolean} True if the series is stated as 'Completed' and False otherwise.
+ */
+export function getMangaStatus(response, url) {
+  const imageDom = response.evaluate('//div[@id="rightContent"]/div[@class="rightBox"]', response, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null);
+  if (!imageDom.singleNodeValue) throw FoxyError(ErrorCode.NO_MANGA_STATUS, url);
+
+  const statusStr = /Status\s(\w+)/.exec(imageDom.singleNodeValue.textContent);
+  if (!statusStr) throw FoxyError(ErrorCode.NO_MANGA_STATUS, url);
+
+  return (statusStr[1] === 'Completed');
 }
 
 /**
@@ -114,10 +126,11 @@ export function getMangaInfo(url) {
     const titleDom = response.evaluate('//meta[@property="og:title"]', headerDom, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null);
     if (!titleDom.singleNodeValue) throw FoxyError(ErrorCode.NO_MANGA_NAME, mangaUrl);
 
-    const name = /Read (.+) (?:Manga|Manhwa|[\d.]+) Online/.exec(titleDom.singleNodeValue.getAttribute('content'))[1];
+    const name = /Read (.+) (?:Manga|Manhwa|Manhua|[\d.]+) Online/.exec(titleDom.singleNodeValue.getAttribute('content'))[1];
 
-    // Get image URL from response
-    const imageUrl = getMangaCover(response);
+    // Get image URL and status from response
+    const imageUrl = getMangaCover(response, mangaUrl);
+    const status = getMangaStatus(response, mangaUrl);
 
     // Get chapter list and return
     try {
@@ -131,6 +144,7 @@ export function getMangaInfo(url) {
         reference: mangaReference,
         url: mangaUrl,
         cover: imageUrl,
+        status,
         last_update: new Date(),
         chapter_list: chapterList,
       };
@@ -164,4 +178,28 @@ export async function updateChapters(manga) {
   } catch (err) {
     throw err;
   }
+}
+
+/**
+ * Returns the manga information
+ * @param {object} manga The manga to update.
+ * @returns {Promise} A promise which resolves to the manga object retrieved.
+ */
+export async function updateMetadata(manga) {
+  if (!manga || !manga.url) {
+    throw new TypeError(`updateMetadata() argument is invalid: ${JSON.stringify(manga)}`);
+  }
+
+  return HttpFetch(manga.url, async (response) => {
+    if (!response || typeof response !== 'object') {
+      throw FoxyError(ErrorCode.RESPONSE_NOT_HTML, manga.url);
+    }
+
+    const update = {
+      cover: getMangaCover(response, manga.url),
+      status: getMangaStatus(response, manga.url),
+    };
+
+    return update;
+  });
 }