+++ static/js/manage-datatables.js
@@ -20,6 +20,7 @@ table.addEventListener('table-loaded', () => {
       null,
       null,
       { searchable: false, orderable: false },
+      null,
     ],
     dom,
     pagingType: 'first_last_numbers',